add merkle tree tests for demo
